+++ LoginApi.swift
@@ -18,7 +18,7 @@ public struct LoginAPI {
     }
     
     // Retrieves a new access token by using the refresh token.
-    public static func refreshAccessToken(refreshToken: String, clientId: String, grantType: String) -> NetworkRequest<OEXAccessToken> {
+    public static func requestTokenWithRefreshToken(refreshToken: String, clientId: String, grantType: String) -> NetworkRequest<OEXAccessToken> {
         let body = ["refresh_token": refreshToken, "client_id": clientId, "grant_type": grantType]
         return NetworkRequest(
             method: .POST,
